pyproject.toml
[tool.poetry]
name = "energy-dashboard-api"
version = "0.1.0"
description = ""
authors = ["tea-tacks <194441018+tea-tacks@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "4.13.3"
dateparser = "1.2.1"
fastapi = ">=0.115.0"
google-api-python-client = "2.165.0"
gspread = "6.2.0"
openai = "1.68.2"
pandas = "2.2.3"
protobuf = "6.30.0"
pydantic = "2.10.6"
python-decouple = "3.8"
python-dateutil = "2.9.0.post0"
requests = "2.32.3"
slowapi = ">=0.1.9"
tenacity = "^9.0.0"
uvicorn = {version = ">=0.34.0", extras = ["standard"]}
coverage = "^7.7.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
playwright = "1.50.0"
pre-commit = "^4.0.1"
pytest = "^8.0.0"
httpx = ">=0.27.0"

[tool.pytest.ini_options]
testpaths = ["api/tests"]
pythonpath = ["api"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
